pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["will <will@loci.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
pydantic = "^2.7.4"
loguru = "^0.7.2"
numpy = "^2.0.0"
cellular-automaton = "^1.0.8"
recordclass = "^0.22.0.2"
pygame = "^2.6.0"
pillow = "^10.4.0"
loci-asset = { path = "../loci-asset", develop = true } 
pywavefront = "^1.3.3"
pygltflib = "^1.16.2"
bpy = ">=3.6.0"
pyrender = "^0.1.45"
mypy = "^1.11.1"
omegaconf = "^2.3.0"
bitstring = "^4.2.3"
pytest = "^8.3.2"
cryptography = "^43.0.0"
scikit-learn = "^1.5.1"
matplotlib = "^3.9.2"
pymongo = "^4.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
